+++ MVC.Service/ICategoryService.cs
@@ -7,13 +7,13 @@ namespace MVC.Service
     public interface ICategoryService
     {
         IEnumerable<Category> GetAllCategories();
-        Category GetCategoryById(long id);
+        Category GetCategoryById(int id);
 
         void InsertCategory(Category category);
 
         void UpdateCategory(Category category);
 
-        void DeleteCategory(long id);
+        void DeleteCategory(int id);
 
     }
 }
